0x01-NoSQL/requirements.txt
anyio==3.7.1
certifi==2023.5.7
colorama==0.4.6
dnspython==2.4.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
mongoengine==0.27.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
